refractor: added utils folder
