+++ Frontend/src/components/Table.jsx
@@ -1,5 +1,6 @@
 import { React } from "react";
 import { useNavigate } from "react-router-dom";
+import { formatPhoneNumber, formatBirthday } from "../utils/stringFormat";
 
 const Table = ({ clients, loading, err }) => {
   const navigate = useNavigate();
@@ -15,26 +16,7 @@ const Table = ({ clients, loading, err }) => {
     console.log(err);
     return <h2>{err}</h2>;
   }
-  const formatPhoneNumber = (phoneNumber) => {
-    let strNum = String(phoneNumber);
 
-    const country = strNum.slice(0, 1);
-    const part1 = strNum.slice(1, 4);
-    const part2 = strNum.slice(4, 7);
-    const part3 = strNum.slice(7, 11);
-
-    return `(+${country}) ${part1} ${part2} ${part3}`;
-  };
-
-  const formatBirthday = (birthday) => {
-    let strNum = String(birthday);
-
-    const day = strNum.slice(8, 10);
-    const month = strNum.slice(5, 7);
-    const year = strNum.slice(0, 4);
-
-    return `${month}-${day}-${year}`;
-  };
   return (
     <table className="table">
       <thead className="text-lg text-white">
